Render a grid with walls and goals colored in
You can set map layouts as in bot_code.pde in elsewhere, and use the
selectors in setup() to choose your layout. Keep in mind that there is
currently no protection for mismatched dimensions between your layout
parameters, i.e. it won't stop you from selecting goals that are outside
the boundaries of the map dimensions. The origin of the axes is in the
top left corner of the render--x grows going right and y grows going
down. For some reason, Processing won't let me change the window size,
so it's stuck at 100x100.